dom/base/test/test_bug1187157.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1187157
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1187157</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1187157">Mozilla Bug 1187157</a>
<form id="a"><input name="b" type="file"/></form>

<script type="text/javascript">
  var obj = new FormData(document.getElementById('a')).get('b');
  ok(obj instanceof File, "This should return a File");
  is(obj.size, 0, "Size should be 0");
  is(obj.name, "", "Name should be the empty string.");
  is(
    obj.type,
    "application/octet-stream",
    "Type should be application/octet-stream"
  );

</script>

</body>
</html>
